split examples into standalone scripts
