+++ pydsfapi/MANIFEST.in
@@ -1,4 +1,4 @@
 include LICENSE
 include README.md
 include setup.py
-include examples.py
+reverse-include examples/*.py
